requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-neo4j
streamlit
faiss-cpu
pypdf
python-dotenv
requests